Refactor code_length method and default total code
