+++ piargus/constants.py
@@ -20,5 +20,3 @@ OPTIMAL = "OPT"
 NETWORK = "NET"
 ROUNDING = "RND"
 TABULAR_ADJUSTMENT = "CTA"
-
-DEFAULT_TOTAL_CODE = "Total"
